+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.2.11
+
+- Update llc dependency
+- Update widget to use `channel.state.unreadCountStream`
+
+## 0.2.10
+
+- Update llc dependency
+- Add `separatorBuilder` to `ChannelListView`
+
 ## 0.2.9+1
 
 - Update llc dependency